+++ index.js
@@ -3,7 +3,7 @@ import GeoTriggeringBuilder from './GeoTriggeringBuilder'
 import TempoBuilder from './TempoBuilder'
 
 const eventEmitter = new NativeEventEmitter(NativeModules.BluedotPointSDK)
-const subscriptionsList = new Set();
+const subscriptionsList = new Set(); // Set of strings - eventNames
 
 const initialize = (projectId, onSucessCallback, onFailCallback) => {
     NativeModules.BluedotPointSDK.initialize(projectId, onSucessCallback, onFailCallback)
@@ -37,6 +37,10 @@ const setCustomEventMetaData = (eventMetaData) => {
     NativeModules.BluedotPointSDK.setCustomEventMetaData(eventMetaData)
 }
 
+const getCustomEventMetaData = () => {
+    return NativeModules.BluedotPointSDK.getCustomEventMetaData()
+}
+
 const setNotificationIdResourceId = (resourceId) => {
     NativeModules.BluedotPointSDK.setNotificationIDResourceID(resourceId)
 }
@@ -45,12 +49,13 @@ const setZoneDisableByApplication = (zoneId, disable) => {
     NativeModules.BluedotPointSDK.setZoneDisableByApplication(zoneId, disable)
 }
 
-const allowsBackgroundLocationUpdates = (enable) => {
-    NativeModules.BluedotPointSDK.allowsBackgroundLocationUpdates(enable)
+const backgroundLocationAccessForWhileUsing = (enable) => {
+    NativeModules.BluedotPointSDK.backgroundLocationAccessForWhileUsing(enable)
 }
 
 const on = (eventName, callback) => {
     eventEmitter.addListener(eventName, callback)
+    subscriptionsList.add(eventName)
 }
 
 const unsubscribe = (eventName) => {
@@ -59,7 +64,7 @@ const unsubscribe = (eventName) => {
 }
 
 const unsubscribeAll = () => {
-    subscriptionsList.forEach(event => eventEmitter.removeAllListeners(event))
+    subscriptionsList.forEach(eventName => eventEmitter.removeAllListeners(eventName))
     subscriptionsList.clear()
 }
 
@@ -80,9 +85,9 @@ const BluedotPointSDK = {
     unsubscribe,
     unsubscribeAll,
     setCustomEventMetaData,
+    getCustomEventMetaData,
     setNotificationIdResourceId,
     getInstallRef,
-    // New APIs
     initialize,
     isInitialized,
     reset,
@@ -95,7 +100,7 @@ const BluedotPointSDK = {
     getSdkVersion,
     getZonesAndFences,
     setZoneDisableByApplication,
-    allowsBackgroundLocationUpdates
+    backgroundLocationAccessForWhileUsing
 }
 
 export default BluedotPointSDK