+++ ios/BluedotPointSDK.m
@@ -11,10 +11,7 @@ RCT_EXPORT_MODULE()
 - (instancetype)init {
     self = [super init];
     if (self) {
-        
-        // Deprecated
-        BDLocationManager.instance.locationDelegate = self;
-        
+
         // New APIs
         BDLocationManager.instance.tempoTrackingDelegate = self;
         BDLocationManager.instance.geoTriggeringEventDelegate = self;
@@ -27,17 +24,17 @@ RCT_EXPORT_MODULE()
 }
 
 RCT_EXPORT_METHOD(initialize:(NSString *)projectId
-    initializationSuccessful:(RCTResponseSenderBlock)initializationSuccessfulCallback
-    initializationFailed: (RCTResponseSenderBlock)initializationFailedCallback)
+                  initializationSuccessful:(RCTResponseSenderBlock)initializationSuccessfulCallback
+                  initializationFailed: (RCTResponseSenderBlock)initializationFailedCallback)
 {
     [[BDLocationManager instance] initializeWithProjectId: projectId completion:^(NSError * error)
      {
-         if (error != nil) {
-             initializationFailedCallback(@[error.localizedDescription]);
-         } else {
-             initializationSuccessfulCallback(@[]);
-         }
-     }];
+        if (error != nil) {
+            initializationFailedCallback(@[error.localizedDescription]);
+        } else {
+            initializationSuccessfulCallback(@[]);
+        }
+    }];
 }
 
 RCT_REMAP_METHOD(isInitialized,
@@ -51,10 +48,10 @@ RCT_REMAP_METHOD(isInitialized,
 }
 
 RCT_EXPORT_METHOD(reset:(RCTResponseSenderBlock)resetSuccessfulCallback
-    resetFailed:(RCTResponseSenderBlock)resetFailedCallback)
+                  resetFailed:(RCTResponseSenderBlock)resetFailedCallback)
 {
     [[BDLocationManager instance] resetWithCompletion:^(NSError * error)
-    {
+     {
         if (error != nil) {
             resetFailedCallback(@[error.localizedDescription]);
         } else {
@@ -68,7 +65,7 @@ RCT_EXPORT_METHOD(iOSStartGeoTriggering:(RCTResponseSenderBlock)startGeoTriggeri
 {
     
     [[BDLocationManager instance] startGeoTriggeringWithCompletion:^(NSError * error)
-    {
+     {
         if (error != nil) {
             startGeoTriggeringFailedCallback(@[error.localizedDescription]);
         } else {
@@ -84,7 +81,7 @@ RCT_EXPORT_METHOD(iOSStartGeoTriggeringWithAppRestartNotification: (NSString *) 
 {
   dispatch_async(dispatch_get_main_queue(),^(void) {
     [[BDLocationManager instance] startGeoTriggeringWithAppRestartNotificationTitle:notificationTitle notificationButtonText:buttonText completion:^(NSError * error)
-    {
+     {
         if (error != nil) {
             startGeoTriggeringTestFailedCallback(@[error.localizedDescription]);
         } else {
@@ -105,10 +102,10 @@ RCT_REMAP_METHOD(isGeoTriggeringRunning,
 }
 
 RCT_EXPORT_METHOD(stopGeoTriggering:(RCTResponseSenderBlock)stopGeoTriggeringSuccessfulCallback
-    stopGeoTriggeringFailed:(RCTResponseSenderBlock)stopGeoTriggeringFailedCallback)
+                  stopGeoTriggeringFailed:(RCTResponseSenderBlock)stopGeoTriggeringFailedCallback)
 {
     [[BDLocationManager instance] stopGeoTriggeringWithCompletion:^(NSError * error)
-    {
+     {
         if (error != nil) {
             stopGeoTriggeringFailedCallback(@[error.localizedDescription]);
         } else {
@@ -158,6 +155,15 @@ RCT_EXPORT_METHOD(setCustomEventMetaData: (NSDictionary *) eventMetaData)
     [ BDLocationManager.instance setCustomEventMetaData: eventMetaData ];
 }
 
+RCT_REMAP_METHOD(getCustomEventMetaData,
+                 getCustomEventMetaDataResolver: (RCTPromiseResolveBlock)resolve
+                 getCustomEventMetaDataRejecter: (RCTPromiseRejectBlock)reject)
+{
+    resolve([BDLocationManager.instance customEventMetaData]);
+}
+
+
+
 RCT_EXPORT_METHOD(setForegroundNotification: (NSString *) channelId
                   channelName: (NSString *) channelName
                   title: (NSString *) title
@@ -170,13 +176,9 @@ RCT_EXPORT_METHOD(setNotificationIDResourceID: (NSString *) resourceID){
     NSLog( @"Note: setNotificationIDResourceID is applicable to Android only");
 }
 
-RCT_EXPORT_METHOD(notifyPushUpdateWithData: (NSDictionary *) data) {
-    [[ BDLocationManager instance] notifyPushUpdateWithData:data];
-}
-
 /*
-*  This method returns a JavaScript Promise. Resolves the installRef from the BDLocationManager and Rejects and error.
-*/
+ *  This method returns a JavaScript Promise. Resolves the installRef from the BDLocationManager and Rejects and error.
+ */
 RCT_REMAP_METHOD(getInstallRef,
                  getInstallRefWithResolver: (RCTPromiseResolveBlock)resolve
                  rejecter: (RCTPromiseRejectBlock)reject)
@@ -200,9 +202,10 @@ RCT_REMAP_METHOD(getSdkVersion,
 }
 
 RCT_REMAP_METHOD(getZonesAndFences,
-                  getZonesAndFencesResolver: (RCTPromiseResolveBlock)resolve
-                  getZonesAndFencesRejecter:(RCTPromiseRejectBlock)reject)
+                 getZonesAndFencesResolver: (RCTPromiseResolveBlock)resolve
+                 getZonesAndFencesRejecter:(RCTPromiseRejectBlock)reject)
 {
+    #warning Update to match new payload in SDK 16 ( + Android)
     NSSet *zoneInfos = [ BDLocationManager.instance zoneInfos ];
     NSMutableArray  *returnZones = [ NSMutableArray new ];
     
@@ -215,21 +218,21 @@ RCT_REMAP_METHOD(getZonesAndFences,
 }
 
 RCT_EXPORT_METHOD(setZoneDisableByApplication: (NSString *) zoneId
-                   disable:(BOOL)disable)
+                  disable:(BOOL)disable)
 {
     [[BDLocationManager instance] setZone:zoneId disableByApplication:disable ];
 }
 
-RCT_EXPORT_METHOD(allowsBackgroundLocationUpdates: (BOOL) enable)
+RCT_EXPORT_METHOD(backgroundLocationAccessForWhileUsing: (BOOL) enable)
 {
-    BDLocationManager.instance.allowsBackgroundLocationUpdates = enable;
+    BDLocationManager.instance.backgroundLocationAccessForWhileUsing = enable;
 }
 
 /*
-    ANDROID METHODS
+ ANDROID METHODS
  
-    The following list of methods are no-op
-    to keep consistency with the Android implementation
+ The following list of methods are no-op
+ to keep consistency with the Android implementation
  */
 
 RCT_EXPORT_METHOD(androidStartGeoTriggering) {
@@ -253,6 +256,7 @@ RCT_EXPORT_METHOD(androidStartTempoTracking) {
         @"enterZone",
         @"exitZone",
         @"tempoTrackingDidExpire",
+        @"tempoTrackingDidUpdate",
         @"tempoTrackingStoppedWithError",
         @"lowPowerModeDidChange",
         @"locationAuthorizationDidChange",
@@ -260,31 +264,39 @@ RCT_EXPORT_METHOD(androidStartTempoTracking) {
     ];
 }
 
-- (void)didEnterZone: (nonnull BDZoneEntryEvent *) enterEvent {
-    NSDictionary *returnFence = [ self fenceToDict: enterEvent.fence ];
-    NSDictionary *returnZone = [ self zoneToDict: enterEvent.zone ];
-    NSDictionary *returnLocation = [ self locationToDict: enterEvent.location ];
-
-    [self sendEventWithName:@"enterZone" body:@{
-        @"fenceInfo" : returnFence,
-        @"zoneInfo" : returnZone,
-        @"locationInfo" : returnLocation,
-        @"isExitEnabled" : [NSNumber numberWithBool: enterEvent.isExitEnabled],
-        @"customData" : enterEvent.zone.customData != nil ? enterEvent.zone.customData : [NSNull null]
-    }];
+/*
+ *  This method is passed the Zone information utilised by the Bluedot SDK.
+ */
+- (void)didUpdateZoneInfo {
+    [self sendEventWithName:@"zoneInfoUpdate" body:nil];
 }
 
-- (void)didExitZone: (nonnull BDZoneExitEvent *) exitEvent {
-    NSDictionary *returnFence = [ self fenceToDict: exitEvent.fence ];
-    NSDictionary *returnZone = [ self zoneToDict: exitEvent.zone ];
-    NSTimeInterval  unixDate = [ exitEvent.date timeIntervalSince1970 ];
-    
-    [self sendEventWithName:@"exitZone" body:@{
-        @"fenceInfo" : returnFence,
-        @"zoneInfo" : returnZone,
-        @"date" : @(unixDate),
-        @"duration" : @(exitEvent.duration)
-    }];
+- (void)didEnterZone:(nonnull GeoTriggerEvent *)enterEvent {
+    NSDictionary *enterEventMap = @{};
+    @try {
+        NSData *data = [[enterEvent toJson:nil] dataUsingEncoding:NSUTF8StringEncoding];
+        enterEventMap = [NSJSONSerialization JSONObjectWithData:data
+                                                        options:0
+                                                          error:nil];
+    } @catch (NSException *exception) {
+
+    } @finally {
+        [self sendEventWithName:@"enterZone" body:enterEventMap];
+    }
+}
+
+- (void)didExitZone: (nonnull GeoTriggerEvent *)exitEvent {
+    NSDictionary *exitEventMap = @{};
+    @try {
+        NSData *data = [[exitEvent toJson:nil] dataUsingEncoding:NSUTF8StringEncoding];
+        exitEventMap = [NSJSONSerialization JSONObjectWithData:data
+                                                        options:0
+                                                          error:nil];
+    } @catch (NSException *exception) {
+
+    } @finally {
+        [self sendEventWithName:@"exitZone" body:exitEventMap];
+    }
 }
 
 - (void)tempoTrackingDidExpire {
@@ -297,6 +309,20 @@ RCT_EXPORT_METHOD(androidStartTempoTracking) {
     [self sendEventWithName:@"tempoTrackingStoppedWithError" body:@{
         @"error" : error.localizedDescription
     }];
+}
+
+- (void)tempoTrackingDidUpdate:(TempoTrackingUpdate *)tempoUpdate {
+    NSDictionary *tempoUpdateMap = @{};
+    @try {
+        NSData *data = [[tempoUpdate toJson:nil] dataUsingEncoding:NSUTF8StringEncoding];
+        tempoUpdateMap = [NSJSONSerialization JSONObjectWithData:data
+                                                       options:0
+                                                         error:nil];
+    } @catch (NSException *exception) {
+
+    } @finally {
+        [self sendEventWithName:@"tempoTrackingDidUpdate" body:tempoUpdateMap];
+    }
 }
 
 - (void)lowPowerModeDidChange:(bool)isLowPowerMode {
@@ -316,24 +342,6 @@ RCT_EXPORT_METHOD(androidStartTempoTracking) {
     [self sendEventWithName:@"accuracyAuthorizationDidChange" body:@{
         @"previousAccuracyAuthorization": @(previousAccuracyAuthorization),
         @"newAccuracyAuthorization": @(newAccuracyAuthorization)
-    }];
-}
-
-/*
-*  This method is passed the Zone information utilised by the Bluedot SDK.
-*/
-- (void)onZoneInfoUpdate:(NSSet<BDZoneInfo *> *)zoneInfos {
-    NSLog( @"Point sdk updated with %lu zones", (unsigned long)zoneInfos.count );
-        
-    NSMutableArray  *returnZones = [ NSMutableArray new ];
-
-    for( BDZoneInfo *zone in zoneInfos )
-    {
-        [ returnZones addObject: [ self zoneToDict: zone ] ];
-    }
-    
-    [self sendEventWithName:@"zoneInfoUpdate" body:@{
-        @"zoneInfos" : returnZones
     }];
 }