+++ ReleaseNotes.md
@@ -1,5 +1,3 @@
 # Bluedot React Native Plugin release notes
 
-- Handled 1st trigger missing callback issue
-- Integrated with Bluedot iOS SDK 15.6.9
-- Integrated with Bluedot Android SDK 15.6.2
+- Integrated with Bluedot Android and iOS SDK 16.0.0